rename statusUtils to watcherUtils
