+++ src/sidecar/connections/watcherUtils.ts
